make new yaml utils file
